No need to be so heavy-handed on overrides. Also shared_children
Capistrano apears to be a lot less rails-specific than it used to be, so
it's fine to not clobber the entire deploy namespace

Also, there's this thing called "shared_children" that creates dirs in
shared_path for you.

Finally, not everthing is a password.  Probably OK to echo the hostname
and username to the screen